AI_Knowledge/STT_TTS_Evaluation/azure_functions/gladiastt001/requirements.txt
azure-functions
azure-storage-blob
azure-data-tables
requests
python-dotenv
